更改了数据库，加入at回复

+++ src/main/java/com/web/service/UserService.java
@@ -7,5 +7,5 @@ import com.web.entity.User;
 public interface UserService {
 	void save(User user);
 	List<User> findAll();
-	User findUserById(Long id);
+	User findUserById(int id);
 }
